+++ draw.h
@@ -12,12 +12,12 @@ public:
   void wipe_out_msg();
   void wipe_out_prg();
   void dump_debug(const char *msg);
-  void dump_progress(double pct);
+  void dump_progress(float pct);
   void dump_board(Board *b);
   void dump_turn(Board *b);
   void dump_score(Board *b);
   void dump_msg(const char *msg);
-  void dump_EWP(double eB, double eW);
+  void dump_EWP(float eB, float eW);
   void dump_who_won(Board *b);
   void update_screen(int stop, int y, int x, Board *b);
   void pause_auto_play();
